+++ server/src/test/java/org/opensearch/gateway/remote/RemoteClusterStateServiceTests.java
@@ -54,6 +54,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
+import java.util.Optional;
 import java.util.Set;
 import java.util.function.Supplier;
 
@@ -140,22 +141,21 @@ public class RemoteClusterStateServiceTests extends OpenSearchTestCase {
         );
     }
 
-    public void testFailWriteFullMetadataWhenRepositoryNotSet() {
-        final ClusterState clusterState = generateClusterStateWithOneIndex().nodes(nodesWithLocalNodeClusterManager()).build();
+    public void testFailInitializeWhenRepositoryNotSet() {
         doThrow(new RepositoryMissingException("repository missing")).when(repositoriesService).repository("remote_store_repository");
-        assertThrows(RepositoryMissingException.class, () -> remoteClusterStateService.writeFullMetadata(clusterState));
+        assertThrows(RepositoryMissingException.class, () -> remoteClusterStateService.start());
     }
 
     public void testFailWriteFullMetadataWhenNotBlobRepository() {
         final FilterRepository filterRepository = mock(FilterRepository.class);
         when(repositoriesService.repository("remote_store_repository")).thenReturn(filterRepository);
-        final ClusterState clusterState = generateClusterStateWithOneIndex().nodes(nodesWithLocalNodeClusterManager()).build();
-        assertThrows(AssertionError.class, () -> remoteClusterStateService.writeFullMetadata(clusterState));
+        assertThrows(AssertionError.class, () -> remoteClusterStateService.start());
     }
 
     public void testWriteFullMetadataSuccess() throws IOException {
         final ClusterState clusterState = generateClusterStateWithOneIndex().nodes(nodesWithLocalNodeClusterManager()).build();
         mockBlobStoreObjects();
+        remoteClusterStateService.start();
         final ClusterMetadataManifest manifest = remoteClusterStateService.writeFullMetadata(clusterState);
         final UploadedIndexMetadata uploadedIndexMetadata = new UploadedIndexMetadata("test-index", "index-uuid", "metadata-filename");
         List<UploadedIndexMetadata> indices = List.of(uploadedIndexMetadata);
@@ -191,6 +191,7 @@ public class RemoteClusterStateServiceTests extends OpenSearchTestCase {
             return null;
         }).when(container).asyncBlobUpload(writeContextArgumentCaptor.capture(), actionListenerArgumentCaptor.capture());
 
+        remoteClusterStateService.start();
         final ClusterMetadataManifest manifest = remoteClusterStateService.writeFullMetadata(clusterState);
 
         final UploadedIndexMetadata uploadedIndexMetadata = new UploadedIndexMetadata("test-index", "index-uuid", "metadata-filename");
@@ -245,6 +246,7 @@ public class RemoteClusterStateServiceTests extends OpenSearchTestCase {
             return null;
         }).when(container).asyncBlobUpload(any(WriteContext.class), actionListenerArgumentCaptor.capture());
 
+        remoteClusterStateService.start();
         assertThrows(
             RemoteClusterStateService.IndexMetadataTransferException.class,
             () -> remoteClusterStateService.writeFullMetadata(clusterState)
@@ -253,6 +255,7 @@ public class RemoteClusterStateServiceTests extends OpenSearchTestCase {
 
     public void testFailWriteIncrementalMetadataNonClusterManagerNode() throws IOException {
         final ClusterState clusterState = generateClusterStateWithOneIndex().build();
+        remoteClusterStateService.start();
         final ClusterMetadataManifest manifest = remoteClusterStateService.writeIncrementalMetadata(clusterState, clusterState, null);
         Assert.assertThat(manifest, nullValue());
     }
@@ -279,7 +282,7 @@ public class RemoteClusterStateServiceTests extends OpenSearchTestCase {
 
         final ClusterMetadataManifest previousManifest = ClusterMetadataManifest.builder().indices(Collections.emptyList()).build();
 
-        remoteClusterStateService.ensureRepositorySet();
+        remoteClusterStateService.start();
         final ClusterMetadataManifest manifest = remoteClusterStateService.writeIncrementalMetadata(
             previousClusterState,
             clusterState,
@@ -319,7 +322,7 @@ public class RemoteClusterStateServiceTests extends OpenSearchTestCase {
             )
         ).thenThrow(IOException.class);
 
-        remoteClusterStateService.ensureRepositorySet();
+        remoteClusterStateService.start();
         Exception e = assertThrows(
             IllegalStateException.class,
             () -> remoteClusterStateService.getLatestClusterMetadataManifest(
@@ -342,15 +345,12 @@ public class RemoteClusterStateServiceTests extends OpenSearchTestCase {
             )
         ).thenReturn(List.of());
 
-        remoteClusterStateService.ensureRepositorySet();
-        Exception e = assertThrows(
-            IllegalStateException.class,
-            () -> remoteClusterStateService.getLatestClusterMetadataManifest(
-                clusterState.getClusterName().value(),
-                clusterState.metadata().clusterUUID()
-            )
+        remoteClusterStateService.start();
+        Optional<ClusterMetadataManifest> manifest = remoteClusterStateService.getLatestClusterMetadataManifest(
+            clusterState.getClusterName().value(),
+            clusterState.metadata().clusterUUID()
         );
-        assertEquals(e.getMessage(), "Remote Cluster State not found - " + clusterState.metadata().clusterUUID());
+        assertEquals(manifest, Optional.empty());
     }
 
     public void testReadLatestMetadataManifestFailedManifestFileRemoveAfterFetchInRemote() throws IOException {
@@ -367,7 +367,7 @@ public class RemoteClusterStateServiceTests extends OpenSearchTestCase {
         ).thenReturn(Arrays.asList(blobMetadata));
         when(blobContainer.readBlob("manifestFileName")).thenThrow(FileNotFoundException.class);
 
-        remoteClusterStateService.ensureRepositorySet();
+        remoteClusterStateService.start();
         Exception e = assertThrows(
             IllegalStateException.class,
             () -> remoteClusterStateService.getLatestClusterMetadataManifest(
@@ -394,7 +394,7 @@ public class RemoteClusterStateServiceTests extends OpenSearchTestCase {
         BlobContainer blobContainer = mockBlobStoreObjects();
         mockBlobContainer(blobContainer, expectedManifest, Map.of());
 
-        remoteClusterStateService.ensureRepositorySet();
+        remoteClusterStateService.start();
         assertEquals(
             remoteClusterStateService.getLatestIndexMetadata(clusterState.getClusterName().value(), clusterState.metadata().clusterUUID())
                 .size(),
@@ -421,7 +421,7 @@ public class RemoteClusterStateServiceTests extends OpenSearchTestCase {
         mockBlobContainer(blobContainer, expectedManifest, Map.of());
         when(blobContainer.readBlob(uploadedIndexMetadata.getUploadedFilename() + ".dat")).thenThrow(FileNotFoundException.class);
 
-        remoteClusterStateService.ensureRepositorySet();
+        remoteClusterStateService.start();
         Exception e = assertThrows(
             IllegalStateException.class,
             () -> remoteClusterStateService.getLatestIndexMetadata(
@@ -449,11 +449,11 @@ public class RemoteClusterStateServiceTests extends OpenSearchTestCase {
             .build();
 
         mockBlobContainer(mockBlobStoreObjects(), expectedManifest, new HashMap<>());
-        remoteClusterStateService.ensureRepositorySet();
+        remoteClusterStateService.start();
         final ClusterMetadataManifest manifest = remoteClusterStateService.getLatestClusterMetadataManifest(
             clusterState.getClusterName().value(),
             clusterState.metadata().clusterUUID()
-        );
+        ).get();
 
         assertThat(manifest.getIndices().size(), is(1));
         assertThat(manifest.getIndices().get(0).getIndexName(), is(uploadedIndexMetadata.getIndexName()));
@@ -467,7 +467,7 @@ public class RemoteClusterStateServiceTests extends OpenSearchTestCase {
 
     public void testReadLatestIndexMetadataSuccess() throws IOException {
         final ClusterState clusterState = generateClusterStateWithOneIndex().nodes(nodesWithLocalNodeClusterManager()).build();
-        remoteClusterStateService.ensureRepositorySet();
+        remoteClusterStateService.start();
 
         final Index index = new Index("test-index", "index-uuid");
         String fileName = "metadata-" + index.getUUID();
@@ -509,7 +509,7 @@ public class RemoteClusterStateServiceTests extends OpenSearchTestCase {
     public void testMarkLastStateAsCommittedSuccess() throws IOException {
         final ClusterState clusterState = generateClusterStateWithOneIndex().nodes(nodesWithLocalNodeClusterManager()).build();
         mockBlobStoreObjects();
-        remoteClusterStateService.ensureRepositorySet();
+        remoteClusterStateService.start();
         final UploadedIndexMetadata uploadedIndexMetadata = new UploadedIndexMetadata("test-index", "index-uuid", "metadata-filename");
         List<UploadedIndexMetadata> indices = List.of(uploadedIndexMetadata);
         final ClusterMetadataManifest previousManifest = ClusterMetadataManifest.builder().indices(indices).build();
@@ -547,8 +547,8 @@ public class RemoteClusterStateServiceTests extends OpenSearchTestCase {
         );
         mockObjectsForGettingPreviousClusterUUID(clusterUUIDsPointers);
 
-        remoteClusterStateService.ensureRepositorySet();
-        String previousClusterUUID = remoteClusterStateService.getLatestClusterUUID("test-cluster");
+        remoteClusterStateService.start();
+        String previousClusterUUID = remoteClusterStateService.getLastKnownUUIDFromRemote("test-cluster");
         assertThat(previousClusterUUID, equalTo("cluster-uuid3"));
     }
 
@@ -563,8 +563,8 @@ public class RemoteClusterStateServiceTests extends OpenSearchTestCase {
         );
         mockObjectsForGettingPreviousClusterUUID(clusterUUIDsPointers);
 
-        remoteClusterStateService.ensureRepositorySet();
-        assertThrows(IllegalStateException.class, () -> remoteClusterStateService.getLatestClusterUUID("test-cluster"));
+        remoteClusterStateService.start();
+        assertThrows(IllegalStateException.class, () -> remoteClusterStateService.getLastKnownUUIDFromRemote("test-cluster"));
     }
 
     private void mockObjectsForGettingPreviousClusterUUID(Map<String, String> clusterUUIDsPointers) throws IOException {